pyproject.toml
[project]
name = "reprod-llm-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "black>=25.12.0",
    "datasets>=4.4.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mecab-ko>=1.0.2",
    "pip>=25.3",
    "python-mecab-ko>=1.3.7",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
